+++ lib/Clarity/Clarity.ml
@@ -1,6 +1,10 @@
 (* This is free and unencumbered software released into the public domain. *)
 
 #include "grammar.ml"
+
+let int_literal z = IntLiteral (Big_int.big_int_of_int z)
+let uint_literal n = UintLiteral (Big_int.big_int_of_int n)
+
 #include "compare.ml"
 #include "print.ml"
 #include "sexp.ml"