+++ lib/Clarity/Clarity.mli
@@ -1,6 +1,10 @@
 (* This is free and unencumbered software released into the public domain. *)
 
 #include "grammar.ml"
+
+val int_literal : int -> literal
+val uint_literal : int -> literal
+
 #include "compare.mli"
 #include "sexp.mli"
 #include "library.mli"